Extracts RightOfWayStateProvider into its own file For consistency with most other files in maliput/api where the file name matches the class name.
